rollup config, resolve circular imports sorta
