+++ extension/data/tbui.js
@@ -4,6 +4,9 @@ import * as TBApi from './tbapi.js';
 import * as TBHelpers from './tbhelpers.js';
 import * as TBCore from './tbcore.js';
 
+import {icons} from './tbconstants';
+export {icons};
+
 const logger = TBLog('TBui');
 const $body = $('body');
 
@@ -27,47 +30,7 @@ let contextMenuClick = false;
     contextMenuClick = await TBStorage.getSettingAsync('GenSettings', 'contextMenuClick', false);
 })();
 
-/**
- * Material design icons mapped to toolbox names through their hexcode.
- *
- * Usage `<div class="tb-icon">${icons.NAME}</div>`
- * @constant {object}
- */
-export const icons = {
-    add: '&#xe145;', // add
-    addBox: '&#xe146;', // add_box
-    addCircle: '&#xe148;', // add_circle_outline
-    arrowLeft: '&#xe314;', // keyboard_arrow_left
-    arrowRight: '&#xe315;', // keyboard_arrow_right
-    close: '&#xe5cd;', // close
-    comments: '&#xe0b7;', // chat
-    delete: '&#xe872;', // delete
-    dotMenu: '&#xe5d4;', // more_vert
-    edit: '&#xe3c9;', // edit
-    help: '&#xe8fd;', // help_outline
-    history: '&#xe889;', // history
-    list: '&#xe896;', // list
-    modlog: '&#xe3ec;', // grid_on
-    modqueue: '&#xe8b2;', // report_problem
-    newModmail: '&#xe168;', // move_to_inbox
-    oldModmail: '&#xe156;', // inbox
-    overlay: '&#xe8ea;', // view_array
-    profile: '&#xe853;', // account_circle
-    refresh: '&#xe5d5;', // refresh
-    remove: '&#xe15b;', // remove
-    settings: '&#xe8b8;', // settings
-    sortDown: '&#xe5db;', // arrow_downward
-    sortUp: '&#xe5d8;', // arrow_upward
-    subTraffic: '&#xe6e1;', // show_chart
-    tbConsole: '&#xe868;', // bug_report
-    tbReload: '&#xe86a;', // cached
-    tbSettingLink: '&#xe157;', // link
-    tbSubConfig: '&#xe869;', // build
-    unmoderated: '&#xe417;', // remove_red_eye
-    userInbox: '&#xe0be;', // email
-    usernote: '&#xe06f;', // note
-};
-    // Icons NOTE: string line length is ALWAYS 152 chars
+// Icons NOTE: string line length is ALWAYS 152 chars
 
 export const logo64 = `iVBORw0KGgoAAAANSUhEUgAAAEAAAABACAYAAACqaXHeAAAABGdBTUEAALGPC/xhBQAAACBjSFJNAAB6JgAAgIQAAPoAAACA6AAAdTAAAOpgAAA6mAAAF3CculE8AAAACXBIWXMAAAsRAAALEQF/ZF+R
                     AAAAGHRFWHRTb2Z0d2FyZQBwYWludC5uZXQgNC4wLjVlhTJlAAAD1ElEQVR4Xu1ZS0hUURg+WY0tNFy0qCiCGpoaC8fXqBEZPWfRsgdRtCgkKBfRIqpFmZugAisLd9YiQsw2thGtsDYVQURBSRQ9FxER
@@ -1984,6 +1947,16 @@ $body.on('click', '.tb-self-expando-button', function () {
         $this.attr('data-state', 'collapsed');
     }
 });
+
+// Misc.
+
+/** Reloads the extension, then reloads the current window. */
+export function reloadToolbox () {
+    textFeedback('toolbox is reloading', FEEDBACK_POSITIVE, 10000, DISPLAY_BOTTOM);
+    browser.runtime.sendMessage({action: 'tb-reload'}).then(() => {
+        window.location.reload();
+    });
+}
 
 // Utilities
